7. Create a controller, route, and view to show a contact forma nd add it to the main menu of your site
ALSO

Site setup with header, footer, default layout so it doesn't look so janky

AND note that the contact form doesn't do anyting right now -I left the action blank because the assignment doesn't say to add a thankyou message or anything. I might come back and do that but I want to focus on the unit tests.